+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.7.8 - 2020/11/29
 - Support change active database from status bar.
--  Support export view.
+-  Support export view, procedure, function.
 
 ## 2.7.5 - 2020/11/19
 - Migrate status to status bar.